+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cph-ng" extension will be documented in this file.
 
+## [Unreleased]
+
+- **feat**: add default file extension setting for Competitive Companion
+
 ## 0.2.1
 
 - **ci**: build failed in pull requests